+++ astroquery/mast/observations.py
@@ -22,8 +22,9 @@ import astropy.coordinates as coord
 from astropy.table import Table, Row, vstack
 from astroquery import log
 from astroquery.mast.cloud import CloudAccess
+from astroquery.utils import commons
 
-from ..utils import commons, async_to_sync
+from ..utils import async_to_sync
 from ..utils.class_or_instance import class_or_instance
 from ..exceptions import (InvalidQueryError, RemoteServiceError,
                           NoResultsWarning, InputWarning)
@@ -234,7 +235,7 @@ class ObservationsClass(MastQueryWithLogin):
         """
 
         # Put coordinates and radius into consistent format
-        coordinates = commons.parse_coordinates(coordinates)
+        coordinates = commons.parse_coordinates(coordinates, return_frame='icrs')
 
         # if radius is just a number we assume degrees
         radius = coord.Angle(radius, u.deg)
@@ -363,7 +364,7 @@ class ObservationsClass(MastQueryWithLogin):
         """
 
         # build the coordinates string needed by ObservationsClass._caom_filtered_position
-        coordinates = commons.parse_coordinates(coordinates)
+        coordinates = commons.parse_coordinates(coordinates, return_frame='icrs')
 
         # if radius is just a number we assume degrees
         radius = coord.Angle(radius, u.deg)